Introduce an example Cloud Native Platform sketch
This repository contains a sketch of how a platform builder might lay out the
configuration of their platform in a git repo. The idea is that the repo would
be a series of 'Configurations' - artifacts built and delivered using packages.
See README.md for a little more context. Note that there are also a series of
issues against this repo tracking open questions, and some more context in the
repo's commit history.

Signed-off-by: Nic Cope <negz@rk0n.org>